refactor(server): Consolidate the streaming messages and storage capabilities in the agent RunWithStream method
