+++ server/agent.go
@@ -235,417 +235,3 @@ func (a *OpenAICompatibleAgentImpl) Run(ctx context.Context, messages []types.Me
 
 	return nil, fmt.Errorf("maximum iterations (%d) reached without final response", maxIterations)
 }
-
-// RunWithStream processes a conversation and returns a streaming response with iterative tool calling support
-func (a *OpenAICompatibleAgentImpl) RunWithStream(ctx context.Context, messages []types.Message) (<-chan *types.Message, error) {
-	if a.llmClient == nil {
-		return nil, fmt.Errorf("no LLM client configured for agent")
-	}
-
-	var tools []sdk.ChatCompletionTool
-	if a.toolBox != nil {
-		tools = a.toolBox.GetTools()
-	}
-
-	outputChan := make(chan *types.Message, 100)
-
-	go func() {
-		defer close(outputChan)
-
-		currentMessages := make([]types.Message, len(messages))
-		copy(currentMessages, messages)
-
-		for iteration := 1; iteration <= a.config.MaxChatCompletionIterations; iteration++ {
-			a.logger.Debug("starting streaming iteration",
-				zap.Int("iteration", iteration),
-				zap.Int("message_count", len(currentMessages)))
-
-			sdkMessages, err := a.converter.ConvertToSDK(currentMessages)
-			if err != nil {
-				a.logger.Error("failed to convert messages to SDK format", zap.Error(err))
-				return
-			}
-
-			if a.config != nil && a.config.SystemPrompt != "" {
-				systemMessage := sdk.Message{
-					Role:    sdk.System,
-					Content: a.config.SystemPrompt,
-				}
-				sdkMessages = append([]sdk.Message{systemMessage}, sdkMessages...)
-			}
-
-			streamResponseChan, streamErrorChan := a.llmClient.CreateStreamingChatCompletion(ctx, sdkMessages, tools...)
-
-			toolCallsExecuted, assistantMessage, toolResultMessages := a.processStreamIteration(ctx, iteration, streamResponseChan, streamErrorChan, outputChan)
-
-			if assistantMessage != nil {
-				currentMessages = append(currentMessages, *assistantMessage)
-			}
-
-			currentMessages = append(currentMessages, toolResultMessages...)
-
-			if len(toolResultMessages) > 0 {
-				lastToolMessage := toolResultMessages[len(toolResultMessages)-1]
-				if lastToolMessage.Kind == "input_required" {
-					a.logger.Debug("streaming completed - input required from user",
-						zap.Int("iteration", iteration),
-						zap.Int("final_message_count", len(currentMessages)))
-					return
-				}
-			}
-
-			if !toolCallsExecuted {
-				a.logger.Debug("streaming completed - no tool calls executed",
-					zap.Int("iteration", iteration),
-					zap.Int("final_message_count", len(currentMessages)))
-				return
-			}
-
-			a.logger.Debug("tool calls executed, continuing to next iteration",
-				zap.Int("iteration", iteration),
-				zap.Int("message_count", len(currentMessages)))
-		}
-
-		a.logger.Warn("max streaming iterations reached", zap.Int("max_iterations", a.config.MaxChatCompletionIterations))
-	}()
-
-	return outputChan, nil
-}
-
-// processStreamIteration processes a single streaming iteration and returns whether tool calls were executed
-func (a *OpenAICompatibleAgentImpl) processStreamIteration(
-	ctx context.Context,
-	iteration int,
-	streamResponseChan <-chan *sdk.CreateChatCompletionStreamResponse,
-	streamErrorChan <-chan error,
-	outputChan chan<- *types.Message,
-) (toolCallsExecuted bool, assistantMessage *types.Message, toolResultMessages []types.Message) {
-	var fullContent string
-	toolCallAccumulator := make(map[int]*sdk.ChatCompletionMessageToolCall)
-	toolResultMessages = make([]types.Message, 0)
-
-	for {
-		select {
-		case <-ctx.Done():
-			return false, nil, nil
-		case streamErr := <-streamErrorChan:
-			if streamErr != nil {
-				a.logger.Error("streaming failed", zap.Error(streamErr))
-			}
-			return false, nil, nil
-		case streamResp, ok := <-streamResponseChan:
-			if !ok {
-				break
-			}
-
-			if streamResp == nil || len(streamResp.Choices) == 0 {
-				continue
-			}
-
-			choice := streamResp.Choices[0]
-
-			if choice.Delta.Content != "" {
-				fullContent += choice.Delta.Content
-
-				chunkMessage := &types.Message{
-					Kind:      "message",
-					MessageID: fmt.Sprintf("chunk-%d-%d", iteration, len(fullContent)),
-					Role:      "assistant",
-					Parts: []types.Part{
-						map[string]any{
-							"kind": "text",
-							"text": choice.Delta.Content,
-						},
-					},
-				}
-
-				select {
-				case outputChan <- chunkMessage:
-				case <-ctx.Done():
-					return false, nil, nil
-				}
-			}
-
-			for _, toolCallChunk := range choice.Delta.ToolCalls {
-				if toolCallAccumulator[toolCallChunk.Index] == nil {
-					toolCallAccumulator[toolCallChunk.Index] = &sdk.ChatCompletionMessageToolCall{
-						Type:     "function",
-						Function: sdk.ChatCompletionMessageToolCallFunction{},
-					}
-				}
-
-				toolCall := toolCallAccumulator[toolCallChunk.Index]
-				if toolCallChunk.ID != "" {
-					toolCall.Id = toolCallChunk.ID
-				}
-				if toolCallChunk.Function.Name != "" {
-					toolCall.Function.Name = toolCallChunk.Function.Name
-				}
-				if toolCallChunk.Function.Arguments != "" {
-					toolCall.Function.Arguments += toolCallChunk.Function.Arguments
-				}
-			}
-
-			if choice.FinishReason != "" {
-				assistantMessage = &types.Message{
-					Kind:      "message",
-					MessageID: fmt.Sprintf("assistant-stream-%d", iteration),
-					Role:      "assistant",
-					Parts:     make([]types.Part, 0),
-				}
-
-				if fullContent != "" {
-					assistantMessage.Parts = append(assistantMessage.Parts, map[string]any{
-						"kind": "text",
-						"text": fullContent,
-					})
-				}
-
-				if len(toolCallAccumulator) > 0 && a.toolBox != nil {
-					toolCalls := make([]sdk.ChatCompletionMessageToolCall, 0, len(toolCallAccumulator))
-					for _, toolCall := range toolCallAccumulator {
-						toolCalls = append(toolCalls, *toolCall)
-					}
-
-					assistantMessage.Parts = append(assistantMessage.Parts, map[string]any{
-						"kind": "data",
-						"data": map[string]any{
-							"tool_calls": toolCalls,
-						},
-					})
-
-					select {
-					case outputChan <- assistantMessage:
-					case <-ctx.Done():
-						return false, nil, nil
-					}
-
-					toolResultMessages = a.executeToolCallsWithEvents(ctx, toolCalls, outputChan)
-
-					return true, assistantMessage, toolResultMessages
-				}
-
-				select {
-				case outputChan <- assistantMessage:
-				case <-ctx.Done():
-				}
-
-				return false, assistantMessage, toolResultMessages
-			}
-		}
-	}
-}
-
-// executeToolCallsWithEvents executes tool calls and emits events, returning tool result messages
-func (a *OpenAICompatibleAgentImpl) executeToolCallsWithEvents(ctx context.Context, toolCalls []sdk.ChatCompletionMessageToolCall, outputChan chan<- *types.Message) []types.Message {
-	toolResultMessages := make([]types.Message, 0)
-
-	for _, toolCall := range toolCalls {
-		if toolCall.Function.Name == "" || toolCall.Id == "" {
-			continue
-		}
-
-		startEvent := &types.Message{
-			Kind:      "message",
-			MessageID: fmt.Sprintf("tool-start-%s", toolCall.Id),
-			Role:      "assistant",
-			Parts: []types.Part{
-				map[string]any{
-					"kind": "data",
-					"data": map[string]any{
-						"tool_name": toolCall.Function.Name,
-						"status":    "started",
-					},
-				},
-			},
-		}
-
-		select {
-		case outputChan <- startEvent:
-		case <-ctx.Done():
-			return toolResultMessages
-		}
-
-		var args map[string]any
-		var result string
-		var toolErr error
-
-		if err := json.Unmarshal([]byte(toolCall.Function.Arguments), &args); err != nil {
-			a.logger.Error("failed to parse tool arguments", zap.String("tool", toolCall.Function.Name), zap.Error(err))
-			result = fmt.Sprintf("Error parsing tool arguments: %s", err.Error())
-			toolErr = err
-
-			failedEvent := &types.Message{
-				Kind:      "message",
-				MessageID: fmt.Sprintf("tool-failed-%s", toolCall.Id),
-				Role:      "assistant",
-				Parts: []types.Part{
-					map[string]any{
-						"kind": "data",
-						"data": map[string]any{
-							"tool_name": toolCall.Function.Name,
-							"status":    "failed",
-						},
-					},
-				},
-			}
-
-			select {
-			case outputChan <- failedEvent:
-			case <-ctx.Done():
-			}
-		} else {
-			// Special handling for input_required tool in streaming mode
-			if toolCall.Function.Name == "input_required" {
-				a.logger.Debug("input_required tool called in streaming mode",
-					zap.String("tool_call_id", toolCall.Id),
-					zap.String("message", toolCall.Function.Arguments))
-
-				// Execute the tool to get the message but then send special input_required message
-				result, toolErr = a.toolBox.ExecuteTool(ctx, toolCall.Function.Name, args)
-
-				completedEvent := &types.Message{
-					Kind:      "message",
-					MessageID: fmt.Sprintf("tool-completed-%s", toolCall.Id),
-					Role:      "assistant",
-					Parts: []types.Part{
-						map[string]any{
-							"kind": "data",
-							"data": map[string]any{
-								"tool_name": toolCall.Function.Name,
-								"status":    "completed",
-							},
-						},
-					},
-				}
-
-				select {
-				case outputChan <- completedEvent:
-				case <-ctx.Done():
-					return toolResultMessages
-				}
-
-				// Add the tool result message
-				toolResultMessage := types.Message{
-					Kind:      "message",
-					MessageID: fmt.Sprintf("tool-result-%s", toolCall.Id),
-					Role:      "tool",
-					Parts: []types.Part{
-						map[string]any{
-							"kind": "data",
-							"data": map[string]any{
-								"tool_call_id": toolCall.Id,
-								"result":       result,
-								"error":        toolErr != nil,
-							},
-						},
-					},
-				}
-
-				select {
-				case outputChan <- &toolResultMessage:
-				case <-ctx.Done():
-					return toolResultMessages
-				}
-
-				toolResultMessages = append(toolResultMessages, toolResultMessage)
-
-				// Now send the special input_required message that will cause task to pause
-				inputMessage := args["message"].(string)
-				inputRequiredMessage := &types.Message{
-					Kind:      "input_required",
-					MessageID: fmt.Sprintf("input-required-%s", toolCall.Id),
-					Role:      "assistant",
-					Parts: []types.Part{
-						map[string]any{
-							"kind": "text",
-							"text": inputMessage,
-						},
-					},
-				}
-
-				select {
-				case outputChan <- inputRequiredMessage:
-				case <-ctx.Done():
-				}
-
-				toolResultMessages = append(toolResultMessages, *inputRequiredMessage)
-
-				return toolResultMessages
-			}
-
-			result, toolErr = a.toolBox.ExecuteTool(ctx, toolCall.Function.Name, args)
-			if toolErr != nil {
-				a.logger.Error("failed to execute tool", zap.String("tool", toolCall.Function.Name), zap.Error(toolErr))
-				result = fmt.Sprintf("Tool execution failed: %s", toolErr.Error())
-
-				failedEvent := &types.Message{
-					Kind:      "message",
-					MessageID: fmt.Sprintf("tool-failed-%s", toolCall.Id),
-					Role:      "assistant",
-					Parts: []types.Part{
-						map[string]any{
-							"kind": "data",
-							"data": map[string]any{
-								"tool_name": toolCall.Function.Name,
-								"status":    "failed",
-							},
-						},
-					},
-				}
-
-				select {
-				case outputChan <- failedEvent:
-				case <-ctx.Done():
-				}
-			} else {
-				completedEvent := &types.Message{
-					Kind:      "message",
-					MessageID: fmt.Sprintf("tool-completed-%s", toolCall.Id),
-					Role:      "assistant",
-					Parts: []types.Part{
-						map[string]any{
-							"kind": "data",
-							"data": map[string]any{
-								"tool_name": toolCall.Function.Name,
-								"status":    "completed",
-							},
-						},
-					},
-				}
-
-				select {
-				case outputChan <- completedEvent:
-				case <-ctx.Done():
-					return toolResultMessages
-				}
-			}
-		}
-
-		toolResultMessage := types.Message{
-			Kind:      "message",
-			MessageID: fmt.Sprintf("tool-result-%s", toolCall.Id),
-			Role:      "tool",
-			Parts: []types.Part{
-				map[string]any{
-					"kind": "data",
-					"data": map[string]any{
-						"tool_call_id": toolCall.Id,
-						"result":       result,
-						"error":        toolErr != nil,
-					},
-				},
-			},
-		}
-
-		select {
-		case outputChan <- &toolResultMessage:
-		case <-ctx.Done():
-			return toolResultMessages
-		}
-
-		toolResultMessages = append(toolResultMessages, toolResultMessage)
-	}
-
-	return toolResultMessages
-}
